--- helpers/isAllBasketChecked.ts
@@ -1,5 +1,0 @@
-import { TBasketItem } from '@/store/basket/type';
-
-export const isAllBasketChecked = (basketItems: TBasketItem[]): boolean => {
-  return basketItems.every((elem) => elem.checked);
-};
